+++ NEWS.md
@@ -5,6 +5,8 @@
 * Fixed issue in `get_data()` which did not correctly backtransform to original
   data when terms had log-transformations such as `log(1 + x)` or `log(x + 1)`.
 
+* Fixed CRAN check issues.
+
 # insight 0.18.3
 
 ## New functions